BlTktExamples>>terminateTerminableUnresponsiveValuableCommandInRunningState bug fix
Fixes: https://github.com/feenkcom/gtoolkit/issues/530

The test previously:

    Created a BlTktCommandTaskExecution
    Created a BlTktTerminableCommand and added it to the executor.
    Created a new process (which was never #resumed) and put it in the executor
    Terminated the command.

This was OK while the termination code just injected the exception code in to the process, because the process was never resumed, so the code wasn't executed.

Now that the #resume is sent, the process is run, but since hasn't been previously there is no call stack and no exception handler, so the process terminates unexpectedly and the VM exits.

What needs to be done is for the task executor to be set up properly (enough) and the process run so that the exception handler catches it and terminates the process gracefully.