Avoid `CollectionIsEmpty: a ProcessList() is empty` exceptions
The exception is triggered in BlHostPulseLoop>>wait when it checks the
priority of the nextReadyProcess. The problem is that nextReadyProcess
checks that the list is not empty and then tries to retrieve the entry,
however the process can be interrupted by a higher priority process
(which also typically moves the process to the back of the queue), which
empties the list between the check and the retrieval.

Fixes: feenkcom/gtoolkit#1635