Rename hierarchy of 4 *Baseline classes
They are now *BaselineMeasurer but name is shortened.
Added deprecation transform rule and keeped old class names with
isDeprecated true.

It was confussing that in mayority of cases, #baseline accessors refer 
to Point locations while a only a couple of the #baseline accessors
were in fact setting the strategy to calculate the baseline location.
So, I chose measurer as it's a word already used on this respect in Bloc.