Refactor visualBounds, and move many methods to Bloc-Sparta.
This change involves multiple selectors but the key method (and the entry
point) is BlElement>>visualBounds.

I'm structuring it similar to measureTextParagraph: and containsPoint:, methods
that require a similar solution in a future: an element having to answer
something that depends on a rendering canvas but the canvas is not in the
execution context.