Fix alignment singleton initialization
The tricky situation is that BlElementAlignment defines a number of class-instance-variables and
subclasses override them with specific values.

I extracted the initialization of class-instance-variable from initialize to  initializeSingletons, which can
send initializeSingletons to super without warnings (as it happens in initialize). All subclasses in the
hierarchy need a class-side initialize method that sends initializeSingletons.

I'm sure we can find a replacement for this implementation.