Move deprecated code to deprecated package
This change introduce a Bloc-Deprecated package and moves the deprecated classes and methods to it. 

The purpose is to not polluate the packages with deprecated stuff and only keep them to let the users migrate their code. (I hope this helps to speed up a bit the navigation in P11 also)