Remove Bloc-States
After a discussion with Martin it seems that states are now unused un Bloc and are now in Toplo with a better implementation. So I propose to remove them.

I moved the code in https://github.com/pharo-graphics/Bloc-LegacyCode

Note that this PR depends on https://github.com/pharo-graphics/Bloc/pull/422 because it would have created conflicts otherwise.

Fixes #423